handle dashboard view

+++ app/controllers/api/me/likes_controller.rb
@@ -9,8 +9,16 @@ class Api::Me::LikesController < Api::ApplicationController
     end
 
     def create
-        like_params[:song_ids].each do |song_id|
-            @current_user.history_likes.like.create(song_id: song_id)
+        if like_params[:song_ids].present?
+            like_params[:song_ids].each do |song_id|
+                @current_user.history_likes.like.create(song_id: song_id)
+            end
+        end
+
+        if like_params[:album_ids].present?
+            like_params[:album_ids].each do |album_id|
+                @current_user.history_likes.like.create(album_id: album_id)
+            end
         end
         
         render json: {success: 'Đã lưu yêu thích'}, status: :created
@@ -19,9 +27,17 @@ class Api::Me::LikesController < Api::ApplicationController
     end
 
     def destroys
-        like_params[:song_ids].each do |song_id|
-          like = @current_user.history_likes.like.find_by(song_id: song_id)
-          like.destroy if like
+        if like_params[:song_ids].present?
+            like_params[:song_ids].each do |song_id|
+            like = @current_user.history_likes.like.find_by(song_id: song_id)
+            like.destroy if like
+            end
+        end
+        if like_params[:album_ids].present?
+            like_params[:album_ids].each do |album_id|
+            like = @current_user.history_likes.like.find_by(album_id: album_id)
+            like.destroy if like
+            end
         end
       
         render json: { success: 'Đã xóa yêu thích' }, status: :ok
@@ -32,7 +48,7 @@ class Api::Me::LikesController < Api::ApplicationController
         private
 
         def like_params
-            params.permit(song_ids: [])
+            params.permit(song_ids: [], album_ids: [])
         end
   
 end
